Refactor spare devices creation and increase their count to 10 (#22)
merge

Merge branch 'master' of git.bubblev.org:bubblev/bubble into kris/additional_spare_devices

Fix check of spare devices count against threshold

Refresh alog's VPN users only when needed

Set constants private and final

Merge branch 'master' into kris/additional_spare_devices

# Conflicts:
#	bubble-server/src/main/java/bubble/server/listener/NodeInitializerListener.java

Fix creation of additional spare devices

Merge branch 'master' into kris/additional_spare_devices

Merge branch 'master' into kris/additional_spare_devices

Merge branch 'master' into kris/additional_spare_devices

Do not create spare devices on sage node for new accounts

Build spare devices only when needed

Merge branch 'master' into kris/additional_spare_devices

Refactor spare devices creation and increase their count to 10

Reviewed-on: https://git.bubblev.org/bubblev/bubble/pulls/22

+++ bubble-server/src/main/java/bubble/dao/device/DeviceDAO.java
@@ -7,30 +7,34 @@ package bubble.dao.device;
 import bubble.dao.account.AccountDAO;
 import bubble.dao.account.AccountOwnedEntityDAO;
 import bubble.dao.app.AppDataDAO;
-import bubble.model.account.Account;
-import bubble.model.cloud.BubbleNetwork;
 import bubble.model.device.BubbleDeviceType;
 import bubble.model.device.Device;
 import bubble.server.BubbleConfiguration;
+import lombok.NonNull;
 import bubble.service.cloud.DeviceIdService;
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.criterion.Order;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+import javax.transaction.Transactional;
 import java.io.File;
 import java.util.List;
 
 import static bubble.ApiConstants.HOME_DIR;
 import static bubble.model.device.Device.UNINITIALIZED_DEVICE_LIKE;
 import static bubble.model.device.Device.newUninitializedDevice;
+import static org.cobbzilla.util.daemon.ZillaRuntime.*;
 import static org.cobbzilla.util.io.FileUtil.abs;
 import static org.cobbzilla.util.io.FileUtil.touch;
+import static org.cobbzilla.util.reflect.ReflectionUtil.copy;
 
 @Repository @Slf4j
 public class DeviceDAO extends AccountOwnedEntityDAO<Device> {
 
-    public static final File VPN_REFRESH_USERS_FILE = new File(HOME_DIR, ".algo_refresh_users");
+    private static final File VPN_REFRESH_USERS_FILE = new File(HOME_DIR, ".algo_refresh_users");
+    private static final short SPARE_DEVICES_PER_ACCOUNT_MAX = 10;
+    private static final short SPARE_DEVICES_PER_ACCOUNT_THRESHOLD = 5;
 
     @Autowired private BubbleConfiguration configuration;
     @Autowired private AccountDAO accountDAO;
@@ -54,44 +58,45 @@ public class DeviceDAO extends AccountOwnedEntityDAO<Device> {
         return super.preCreate(device);
     }
 
-    @Override public Device create(Device device) {
-        if (!device.uninitialized()) {
-            final String account = device.getAccount();
-            final String network = device.getNetwork();
-            final List<Device> devices = findByAccountAndUninitialized(account);
+    @Transactional
+    @Override public Device create(@NonNull final Device device) {
+        if (device.uninitialized()) return super.create(device);
+        device.initDeviceType();
 
-            // if we have an uninitialized device, use that, otherwise create a new uninitialized device and use that
-            final Device uninitialized;
-            if (devices.isEmpty()) {
-                uninitialized = ensureSpareDevice(account, network, false);
-            } else {
-                uninitialized = devices.get(0);
-            }
-            if (uninitialized != null) {
-                uninitialized.initialize(device);
-                final Device updated = update(uninitialized);
+        final var accountUuid = device.getAccount();
+        final var uninitializedDevices = findByAccountAndUninitialized(accountUuid);
 
-                // always ensure we have one spare device
-                ensureSpareDevice(account, network, true);
-                return updated;
-            }
+        var newDevicesCreated = false;
+        if (uninitializedDevices.size() <= SPARE_DEVICES_PER_ACCOUNT_THRESHOLD
+                && !configuration.getBean(AccountDAO.class).findByUuid(accountUuid).isRoot()) {
+            newDevicesCreated = ensureAllSpareDevices(accountUuid, device.getNetwork());
         }
-        return super.create(device);
+
+        final Device result;
+        // run the above creation of spare devices in parallel, but if there were no spare devices loaded before that,
+        // create a brand new entry here:
+        if (uninitializedDevices.isEmpty()) {
+            log.info("create: no uninitialized devices for account " + accountUuid);
+            // just create the device now:
+            device.initTotpKey();
+            result = super.create(device);
+            newDevicesCreated = true;
+        } else {
+            final var uninitialized = uninitializedDevices.get(0);
+            copy(uninitialized, device);
+            result = super.update(uninitialized);
+        }
+
+        if (newDevicesCreated) refreshVpnUsers();
+        return result;
     }
 
-    @Override public Device update(Device device) {
-        final Device current = findByUuid(device.getUuid());
-        if (current != null && current.uninitialized()) {
-            device.initTotpKey();
-            if (device.getDeviceType().hasDefaultSecurityLevel()) {
-                device.setSecurityLevel(device.getDeviceType().getDefaultSecurityLevel());
-            }
-        }
-        if (device.getDeviceType() == null || device.getDeviceType() == BubbleDeviceType.uninitialized) {
-            device.setDeviceType(BubbleDeviceType.other);
-        }
-        final Device updated = super.update(device);
-        ensureSpareDevice(device.getAccount(), device.getNetwork(), true);
+    @Override @NonNull public Device update(@NonNull final Device updateRequest) {
+        final var toUpdate = (Device) readOnlySession().get(Device.class, updateRequest.getUuid());
+        if (toUpdate == null) die("Cannot find device to update with uuid: " + updateRequest.getUuid());
+        if (toUpdate.uninitialized()) die("Cannot update special devices: " + updateRequest.getName());
+
+        final var updated = super.update(toUpdate);
         deviceIdService.setDeviceSecurityLevel(updated);
         return updated;
     }
@@ -99,37 +104,31 @@ public class DeviceDAO extends AccountOwnedEntityDAO<Device> {
     @Override public void delete(String uuid) {
         final Device device = findByUuid(uuid);
         if (device != null) {
+            if (device.uninitialized()) die("Cannot delete special device: " + device.getName());
+
             dataDAO.deleteDevice(uuid);
             super.delete(uuid);
-            ensureSpareDevice(device.getAccount(), device.getNetwork(), true);
+            refreshVpnUsers();
         }
     }
 
     @Override public void forceDelete(String uuid) { super.delete(uuid); }
 
-    public Device ensureSpareDevice(String accountUuid, String network, boolean refreshVpnUsers) {
-        final BubbleNetwork thisNetwork = configuration.getThisNetwork();
-        final Account account = accountDAO.findByUuid(accountUuid);
-        if (account == null) {
-            log.warn("ensureSpareDevice: account not found: " + accountUuid);
-            return null;
+    @Transactional
+    public synchronized boolean ensureAllSpareDevices(@NonNull final String account, @NonNull final String network) {
+        final var currentSpareDevices = findByAccountAndUninitialized(account);
+        boolean newDevicesCreated = false;
+        for (var i = currentSpareDevices.size(); i < SPARE_DEVICES_PER_ACCOUNT_MAX; i++) {
+            super.create(newUninitializedDevice(network, account));
+            newDevicesCreated = true;
         }
-        if (account.getHashedPassword().isSpecial()) {
-            log.warn("ensureSpareDevice: not creating spare device for special user: " + accountUuid);
-            return null;
-        }
-        final List<Device> devices = findByAccountAndUninitialized(accountUuid);
-        Device uninitialized;
-        if (devices.isEmpty()) {
-            log.info("ensureSpareDevice: no uninitialized devices for account " + accountUuid + ", creating one");
-            uninitialized = create(newUninitializedDevice(network, accountUuid));
-        } else {
-            uninitialized = devices.get(0);
-        }
-        if (refreshVpnUsers) refreshVpnUsers();
-        return uninitialized;
+        return newDevicesCreated;
     }
 
+    /**
+     * This refresh should be done only if there was real change in device count. The script that is executed afterwards
+     * uses only devices' UUIDs, so this should be called . No need to call this method in any other case.
+     */
     public void refreshVpnUsers() {
         log.info("ensureSpareDevice: refreshing VPN users by touching: "+abs(VPN_REFRESH_USERS_FILE));
         touch(VPN_REFRESH_USERS_FILE);
